requirements.txt
attrs==23.2.0
autopep8==2.0.4
aws-cdk-lib==2.130.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
boto3==1.34.51
botocore==1.34.51
cattrs==23.2.3
constructs==10.3.0
exceptiongroup==1.2.0
importlib_resources==6.1.2
jmespath==1.0.1
jsii==1.94.0
publication==0.0.3
pycodestyle==2.11.1
python-dateutil==2.8.2
s3transfer==0.10.0
six==1.16.0
tomli==2.0.1
typeguard==2.13.3
typing_extensions==4.10.0
urllib3==1.26.18
zipp==3.17.0